+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ lib_ethernet change log
   * ADDED: Tests for SMI
   * ADDED: Optional credit limit for Qav shaper
   * ADDED: Support for running MII sim tests for XS3 architecture
+  * ADDED: Hardware tests for RMII Ethernet MAC
   * CHANGED: SMI re-write. Single port version now functional and documented
   * CHANGED: Moved example PHY drivers to lib_board_support
   * REMOVED: Slicekit based examples because hardware is obsolete